requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
fastapi==0.115.12
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   requests
pydantic==2.11.5
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.0
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
typing-extensions==4.14.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.4.0
    # via requests
